+++ src/router.js
@@ -1,6 +1,7 @@
 import Vue from 'vue'
 import Router from 'vue-router'
-import Home from './views/Home.vue'
+import Mapa from './views/Mapa.vue'
+import Dodaj from './views/Dodaj.vue'
 
 Vue.use(Router)
 
@@ -11,15 +12,12 @@ export default new Router({
     {
       path: '/',
       name: 'home',
-      component: Home
+      component: Mapa
     },
     {
-      path: '/about',
-      name: 'about',
-      // route level code-splitting
-      // this generates a separate chunk (about.[hash].js) for this route
-      // which is lazy-loaded when the route is visited.
-      component: () => import(/* webpackChunkName: "about" */ './views/About.vue')
+      path: '/dodaj',
+      name: 'dodaj',
+      component: Dodaj
     }
   ]
 })
